NeuroScan-AI-main/requirements.txt
streamlit
opencv-python-headless
numpy
pillow
keras
tensorflow            
streamlit-lottie
requests
plotly
reportlab
google-generativeai
